+++ src/views/Students/Todaylesson.tsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react'
-import Popup from '../components/Popup'
-import AddStudent from '../Formy/AddStudent'
+import Popup from '../../components/Popup'
+import AddStudent from '../../Formy/AddStudent'
 import axios from 'axios';
 
 const Students = () => {
@@ -10,7 +10,7 @@ const Students = () => {
   useEffect(() => {
     const fetchStudents = async () => {
       try {
-        const response = await axios.get('http://localhost:8080/api/students');
+        const response = await axios.get('http://localhost:8080/api/students/today_lessons');
         setStudents(response.data);
       } catch (err) {
         console.error('Error fetching students:', err);
@@ -33,7 +33,7 @@ const Students = () => {
 
 <aside id="default-sidebar" className="fixed top-0 left-0 z-40 w-64 h-screen transition-transform -translate-x-full sm:translate-x-0" aria-label="Sidenav">
 <div className="overflow-y-auto py-5 px-3 h-full  border-r  bg-gray-800 border-gray-700">
-<ul className="space-y-2">
+    <ul className="space-y-2">
         <li>
           <Popup buttonText="Add Vehicle">
                 <AddStudent />
@@ -48,7 +48,7 @@ const Students = () => {
         <li>
             <a href="/Todaylesson" className="flex items-center p-2 text-base font-normal text-blue-700 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group">
                 <svg aria-hidden="true" className="w-6 h-6 text-gray-400 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 dark:group-hover:text-white" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M2 10a8 8 0 018-8v8h8a8 8 0 11-16 0z"></path><path d="M12 2.252A8.014 8.014 0 0117.748 8H12V2.252z"></path></svg>
-                <span className="ml-3">today lesson</span>
+                <span className="ml-3">Todaylesson</span>
             </a>
         </li>
